requirements-collector.txt
PyYAML>=6.0
# Optional workflow-generation extras
python-dotenv>=1.0.0
openai>=1.0.0
